move test to aggregate_fuzz
